Backend update

+++ backend/database/migrations/2024_03_30_105403_add_user_id_to_films_table.php
@@ -9,13 +9,17 @@ class AddUserIdToFilmsTable extends Migration
     public function up()
     {
         Schema::table('films', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            // Vérifie si la colonne 'user_id' n'existe pas déjà
+            if (!Schema::hasColumn('films', 'user_id')) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('films', function (Blueprint $table) {
+            // Supprime la contrainte de clé étrangère et la colonne 'user_id'
             $table->dropForeign(['user_id']);
             $table->dropColumn('user_id');
         });
